* commit '8aabb8d9': (268 commits)
  Update to syn-2
  devices: vsock: add connection logging for Windows.
  cros_tracing: fix interface drift.
  devices: virtio: scsi: Raise error when trying to modify a read-only disk
  e2e_tests: scsi: Add test cases for discard operations.
  docs: Add SCSI section.
  base: unix: return Vec<SafeDecriptor> from recv_with_fds()
  base: unix: rework ScmSocket send/recv functions
  Only support cpufreq on arm/aarch64
  base: add IoBuf/IoBufMut mut slice conversions
  vmm_vhost: eliminate locking in SlaveReqHandler and merge traits
  vmm_vhost: simplify Slave
  vmm_vhost: eliminate locking in MasterReqHandler and merge traits
  Don't expose `base::platform` as a public interface
  usb: refactor transfer structs for multi backends
  devices: virtio: scsi: Change the queue size.
  devices: vhost-user: console: clean up arg parsing
  hypervisor/README.md: add qualcomm contacts for gunyah
  vmm_vhost: remove type parameter from PlatformConnection
  vmm_vhost: simplify backend request connection creation
  ...

Bug: 307396704
Test: .

+++ devices/src/virtio/gpu/virtio_gpu.rs
@@ -19,7 +19,7 @@ use base::FromRawDescriptor;
 use base::IntoRawDescriptor;
 use base::Protection;
 use base::SafeDescriptor;
-use data_model::VolatileSlice;
+use base::VolatileSlice;
 use gpu_display::*;
 use libc::c_void;
 use rutabaga_gfx::ResourceCreate3D;
@@ -789,7 +789,7 @@ impl VirtioGpu {
     }
 
     /// If supported, export the fence with the given `fence_id` to a file.
-    pub fn export_fence(&self, fence_id: u32) -> ResourceResponse {
+    pub fn export_fence(&self, fence_id: u64) -> ResourceResponse {
         match self.rutabaga.export_fence(fence_id) {
             Ok(handle) => ResourceResponse::Resource(ResourceInfo::Fence {
                 handle: to_safe_descriptor(handle.os_handle),
